// src/types/action-types.ts
export enum ActionType {
  Add = 'add',
  Delete = 'delete',
  DeleteAll = 'delete-all',
  AddToHead = 'add-to-head',
  AddToTail = 'add-to-tail ',
  DeleteHead = 'delete-head',
  DeleteTail = 'delete-tail',
  AddToHeadPerIndex = 'add-to-head-per-index',
  AddToTailPerIndex = 'add-to-tail-per-index ',
}